Move revocation list related code into new revocation_list module
